[REF] pages: added a pages folder and moved dashboard, login, and registry from the components folder to the new pages folder
