+++ fitness_app/src/pages/Dashboard/styles/UserCaloriesChart.module.css
