+++ fitness_app/src/pages/Dashboard/styles/UserPerformanceChart.module.css
