+++ fitness_app/src/pages/Registry/styles/Registry.module.css
